:sparkles: Feature #22: NavRoute로 경로 관리되도록 수정, 스페이스 화면에서 도형뷰 사라짐
