+++ app/src/main/java/com/umc/edison/ui/navigation/BottomNavigation.kt
@@ -4,6 +4,7 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBar
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.material3.NavigationBarItemDefaults
@@ -13,10 +14,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.umc.edison.R
@@ -24,53 +23,54 @@ import com.umc.edison.ui.theme.EdisonTheme
 import com.umc.edison.ui.theme.Gray100
 import com.umc.edison.ui.theme.Gray400
 import com.umc.edison.ui.theme.Gray800
-import com.umc.edison.ui.theme.Pretendard
 import com.umc.edison.ui.theme.White000
 
 sealed class BottomNavItem(
     val title: Int,
     val icon: Int,
     val selectedIcon: Int,
-    val route: String
+    val route: String? = null
 ) {
     data object MyEdison : BottomNavItem(
         title = R.string.my_edison,
         icon = R.drawable.ic_my_edison,
         selectedIcon = R.drawable.ic_my_edison_selected,
-        route = "MY_EDISON"
+        route = NavRoute.MyEdison.route
     )
 
     data object Space : BottomNavItem(
         title = R.string.space,
         icon = R.drawable.ic_space,
         selectedIcon = R.drawable.ic_space_selected,
-        route = "SPACE"
+        route = NavRoute.Space.route
     )
 
     data object Bubble : BottomNavItem(
         title = R.string.bubble,
         icon = R.drawable.ic_bubble,
         selectedIcon = R.drawable.ic_bubble,
-        route = "BUBBLE"
     )
 
     data object ArtLetter : BottomNavItem(
         title = R.string.artletter,
         icon = R.drawable.ic_art_letter,
         selectedIcon = R.drawable.ic_art_letter_selected,
-        route = "ART_LETTER"
+        route = NavRoute.ArtBoard.route
     )
 
     data object MyPage : BottomNavItem(
         title = R.string.my,
         icon = R.drawable.ic_my,
         selectedIcon = R.drawable.ic_my_selected,
-        route = "MY_PAGE"
+        route = NavRoute.MyPage.route
     )
 }
 
 @Composable
-fun BottomNavigation(navController: NavHostController) {
+fun BottomNavigation(
+    navController: NavHostController,
+    onBubbleClick: () -> Unit = {}
+) {
     val items = listOf(
         BottomNavItem.MyEdison,
         BottomNavItem.Space,
@@ -98,9 +98,9 @@ fun BottomNavigation(navController: NavHostController) {
                             contentDescription = navItem.route
                         )
                     },
-                    selected = currentRoute == navItem.route,
+                    selected = false,
                     onClick = {
-                        navController.navigate(navItem.route)
+                        onBubbleClick()
                     },
                     colors = NavigationBarItemDefaults.colors(
                         selectedIconColor = Gray800,
@@ -120,16 +120,12 @@ fun BottomNavigation(navController: NavHostController) {
                         Text(
                             text = LocalContext.current.getString(navItem.title),
                             color = if (currentRoute == navItem.route) Gray800 else Gray400,
-                            style = TextStyle(
-                                fontFamily = Pretendard,
-                                fontSize = 12.sp,
-                                lineHeight = 14.sp
-                            ) // TODO: 추후 Typography 정리되면 수정하기
+                            style = MaterialTheme.typography.labelSmall
                         )
                     },
                     selected = currentRoute == navItem.route,
                     onClick = {
-                        navController.navigate(navItem.route)
+                        navController.navigate(navItem.route!!)
                     },
                     colors = NavigationBarItemDefaults.colors(
                         selectedIconColor = Gray800,
