+++ app/src/main/java/com/umc/edison/ui/navigation/NavigationGraph.kt
@@ -2,30 +2,40 @@ package com.umc.edison.ui.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
+import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import androidx.navigation.navArgument
 import com.umc.edison.ui.artboard.ArtBoardScreen
 import com.umc.edison.ui.my_edison.MyEdisonScreen
 import com.umc.edison.ui.mypage.MyPageScreen
 import com.umc.edison.ui.space.BubbleSpaceScreen
+import com.umc.edison.ui.space.LabelDetailScreen
 
 @Composable
 fun NavigationGraph(navHostController: NavHostController) {
-    NavHost(navHostController, startDestination = BottomNavItem.MyEdison.route) {
-        composable(BottomNavItem.MyEdison.route) {
-            MyEdisonScreen()
+    NavHost(navHostController, startDestination = NavRoute.MyEdison.route) {
+        // bottom navigation
+        composable(NavRoute.MyEdison.route) {
+            MyEdisonScreen(navHostController)
         }
-        composable(BottomNavItem.Space.route) {
-            BubbleSpaceScreen()
+        composable(NavRoute.Space.route) {
+            BubbleSpaceScreen(navHostController)
         }
-        composable(BottomNavItem.Bubble.route) {
-//            BubbleScreen()
+        composable(NavRoute.ArtBoard.route) {
+            ArtBoardScreen(navHostController)
         }
-        composable(BottomNavItem.ArtLetter.route) {
-            ArtBoardScreen()
+        composable(NavRoute.MyPage.route) {
+            MyPageScreen(navHostController)
         }
-        composable(BottomNavItem.MyPage.route) {
-            MyPageScreen()
+
+        // space screen
+        composable(
+            route = "${NavRoute.SpaceLabel.route}/{id}",
+            arguments = listOf(navArgument("id") { type = NavType.IntType })
+        ) { backStackEntry ->
+            val id = backStackEntry.arguments?.getInt("id")
+            LabelDetailScreen(navHostController, id)
         }
     }
 }
