+++ app/src/main/java/com/umc/edison/ui/space/LabelDetailScreen.kt
@@ -1,0 +1,12 @@
+package com.umc.edison.ui.space
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+
+@Composable
+fun LabelDetailScreen(
+    navHostController: NavHostController,
+    id: Int?
+) {
+
+}
